Implement more "Default" parameters.
Looking at just temoa_island.dat, for our modeling purposes, there are a
number of repeated datum elements in CostFixed and CostMarginal.  This
change allows a modeler to define default fixed and marginal costs per
vintage of a technology, such that as long as that process is alive, it
will cost the same (in nominal dollars).  This still affords the modeler
the ability to override any default if they so choose.

New parameters:

CostFixedVintageDefault
CostMarginalVintageDefault