Update Tech*Split implementation.
In conversation, Joe and I had talked about reworking the logic of the
T*S constraints.  Basically, the change is to only consider one side or
the other of the process for the T*S constraint.  The previous
implementation required the modeler to specify a split for individual
inputs or outputs.  Now it's global to the technology.

This also greatly simplifies the logic.